Fixed printf format errors.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@38022 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ contrib/nvi/ex/ex_print.c
@@ -122,7 +122,8 @@ ex_print(sp, cmdp, fp, tp, flags)
 		 */
 		if (LF_ISSET(E_C_HASH)) {
 			if (from <= 999999) {
-				snprintf(buf, sizeof(buf), "%6ld  ", from);
+				snprintf(buf, sizeof(buf), "%6lu  ",
+				    (u_long)from);
 				p = buf;
 			} else
 				p = "TOOBIG  ";
